feat: autenticação adicionada

+++ server/src/contollers/memoriesController.ts
@@ -8,6 +8,7 @@ async function getAll(request: FastifyRequest, reply: FastifyReply) {
   try {
     const { page, size } = pageableFormatter(request);
     const memories = await prisma.memory.findMany({
+      where: { userId: request.user.sub },
       orderBy: { createdAt: 'asc' },
       skip: size * (page - 1),
       take: size,
@@ -40,6 +41,10 @@ async function getById(request: FastifyRequest, reply: FastifyReply) {
     const id = idExtractor(request);
     const memory = await prisma.memory.findUniqueOrThrow({ where: { id } });
 
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(403).send({ message: 'Forbidden' });
+    }
+
     return reply.send({ data: memory });
   } catch (error: any) {
     errorHandler(reply, error);
@@ -54,7 +59,7 @@ async function create(request: FastifyRequest, reply: FastifyReply) {
         content,
         coverUrl,
         isPublic,
-        userId: 'ea74e73b-4b04-4c46-bae9-243416e2c53d', // TODO: get user id from request
+        userId: request.user.sub,
       },
     });
 
@@ -92,7 +97,13 @@ async function update(request: FastifyRequest, reply: FastifyReply) {
 async function remove(request: FastifyRequest, reply: FastifyReply) {
   try {
     const id = idExtractor(request);
-    const memory = await prisma.memory.delete({ where: { id } });
+
+    let memory = await prisma.memory.findUniqueOrThrow({ where: { id } });
+    if (memory.userId !== request.user.sub) {
+      return reply.status(403).send({ message: 'Forbidden' });
+    }
+
+    memory = await prisma.memory.delete({ where: { id } });
 
     return reply.send({ data: memory });
   } catch (error) {
@@ -100,4 +111,8 @@ async function remove(request: FastifyRequest, reply: FastifyReply) {
   }
 }
 
-export const memoriesController = { getAll, getById, create, update, remove };
+async function preHandler(request: FastifyRequest) {
+  await request.jwtVerify();
+}
+
+export const memoriesController = { getAll, getById, create, update, remove, preHandler };
